Code clean-up and re-organization.
